Check DEFAULT_TIMEOUT (Django 1.6+)
There is a new DEFAULT_TIMEOUT object in Django which is checked when timeout is not set.
It has been introduced in:
https://github.com/django/django/commit/89955cc35f3636684ea6f2a6c9504b35a3050f0f

The problem was that this has been passed to uWSGI because it has not been checked and the uWSGI
cache_update function expects an integer not an object() instace, so it gave a TypeError.

Also check settings.py default values the same way as it be would passed to function call.
(So we need to check every value in _set() which comes from settings.py.)